Changed g.! variable prefixes to g.0 variable prefixes.
The original naming convention only ever worked under IBM's VM/CMS
operating system. Modern rexx allows variables beginning with !
but not with 0 (or any digit). This naming convention ensures that variable
substitution is not accidentally performed. For example, g.!var could be
interpreted as g.77 if !var = 77, but because there is no way to assign
77 to a variable called 0var, g.0var will only ever refer
to the intended item.