Added a showNodeNonRecursive() function and invoked it by default when using IBM z/OS TSO.
Strange but true: the z/OS flavor of Rexx has a fixed recursion
limit of 250 (including nesting of if/select clauses as well as call/function
nesting)! Compared to Linux (Regina Rexx) which has a limit of
approximately 164000, this is appalling. Apparently the Rexx
interpreter on z/OS has not been updated since 1992 when memory
was scarce and expensive. I'm appalled. This is appalling.

So...for TSO the DOM tree is now walked without recursion for z/OS.
It may be slightly quicker because of this, but it is certainly
more difficult to understand than simply using recursion.